Fix disabled mods still being loaded
The format of mod-list.json written on syncMods would incorrectly write
a base mod without an enableds key which is required, causing Factorio
to ignore the file entirerly.  Furthermore the enabled status was not
written causing all mods to be loaded even if they were disabled in the
mod pack.