Ignore techonolgy enabled status for sync
Technologies that were disabled were ignored completely by research sync
due to a missunderstanding of what disabled means.  Disabled does not
mean the tech is somehow inaccessible and completely out of play, it
only means it can't be selected in the technology screen.

Ignore the enabled status of technologies in research_sync to fix a
crash that happened when a mod enables a disabled by default technology
and the player attempts to research it.