Fix debug dump-ws crashing the controller
When ws was updated to version 8 it came with a breaking change where
message data was no longer utf8 decoded from binary to a string.  The
main codepath handling messages was updated to handle this but the ws
dump path was not resulting in the controller crashing when receiving
a message that is then forwarded to the dumper without decoding.

Fix by decoding the message in the dump-ws path.