Refactor config loading into function
Preparation for a command that creates a config. Such command should not
fail when there is no existing config (that is the point after all).
However, other config commands should be able to load the config. Hence,
the handler of config commands needs to be able to load the config if
needed.