Remove "instance" from ConfigLocation
When the config system was ported to TypeScript the "instance" location
was accidentally added to the type of available locations for access
control.  This has never been a valid location as it was deemed too
complicated to work with.  The correct location for instances is
"controller" and "host" as these are the locations the configs are
handled.